tesseract/requirements.txt
apt install imagemagick
apt install tesseract-ocr
apt install libtesseract-dev libleptonica-dev
pip install tesserocr

tesseract 4.0.0-beta.1
